// src/app/Core/Pages/dashboard/Modules/home/models/timeline.ts
export interface TimelineCreatedDate {
  name: string;
  createdDate: Date;
  module:string;
}
export interface TimelineUpdatedDate {
  name: string;
  updatedDate: Date;
  module:string;

}
export interface TimelineStartedDate {
  name: string;
  startedDate: Date;
  module:string;

}
export interface TimelineEndedDate {
  name: string;
  endedDate: Date;
  module:string;

}